+++ backend/backend/handlers/subscription/subscriptionService.py
@@ -28,6 +28,7 @@ main_rest_response = STANDARD_JSON_RESPONSE
 try:
     subscription_table_name = os.environ["SUBSCRIPTIONS_STORAGE_TABLE_NAME"]
     asset_table_name = os.environ["ASSET_STORAGE_TABLE_NAME"]
+    user_table_name = os.environ["USER_STORAGE_TABLE_NAME"]
 except:
     logger.exception("Failed loading environment variables")
     main_rest_response['body'] = json.dumps(
@@ -160,23 +161,25 @@ def delete_sns_subscriptions(asset_id, subscribers, delete_sns=False):
     if not asset_obj.get("snsTopic"):
         logger.error(f"No topic found for asset {asset_id}")
         return
-
-    resp = sns_client.list_subscriptions_by_topic(TopicArn=asset_obj.get("snsTopic"))
-    subscription_arns = [subscription['SubscriptionArn'] for subscription in resp['Subscriptions'] if subscription['Endpoint'] in subscribers]
-
-    for subscription_arn in subscription_arns:
-        if subscription_arn != "PendingConfirmation":
-            sns_client.unsubscribe(SubscriptionArn=subscription_arn)
-
+    
     if delete_sns:
         sns_client.delete_topic(TopicArn=asset_obj.get("snsTopic"))
         asset_table.update_item(
             Key={'databaseId': asset_obj["databaseId"], 'assetId': asset_id},
             UpdateExpression=f"REMOVE snsTopic"
         )
+    else:
+        resp = sns_client.list_subscriptions_by_topic(TopicArn=asset_obj.get("snsTopic"))
+        subscription_arns = [subscription['SubscriptionArn'] for subscription in resp['Subscriptions'] if subscription['Endpoint'] in subscribers]
+
+        for subscription_arn in subscription_arns:
+            if subscription_arn != "PendingConfirmation":
+                sns_client.unsubscribe(SubscriptionArn=subscription_arn)
+
+    return
 
 
-def create_sns_subscriptions(asset_id, subscribers):
+def create_sns_subscriptions(asset_id, emails):
     asset_obj = get_asset(asset_id)
     asset_sns_topic = asset_obj.get("snsTopic")
 
@@ -184,7 +187,7 @@ def create_sns_subscriptions(asset_id, subscribers):
         asset_sns_topic = create_sns_topic(asset_id)
         add_sns_topic_in_asset(asset_id, asset_obj["databaseId"], asset_sns_topic)
 
-    for subscriber in subscribers:
+    for subscriber in emails:
         sns_client.subscribe(
             TopicArn=asset_sns_topic,
             Protocol='email',
@@ -203,10 +206,51 @@ def get_subscription_obj(event_name, entity_name, entity_id):
     return resp.get('Item')
 
 
+def get_userProfile_Email(userId):
+
+    #Try to get user email information
+    user_table = dynamodb.Table(user_table_name)
+    response = user_table.get_item(
+        Key={
+            'userId': userId
+        }
+    )
+
+    #Lookup user profile email and use that. 
+    #If not available or blank, set to userID and validate it's in email format
+    email = None
+    if 'Item' in response:
+        email = response['Item'].get('email','')
+
+    if not email or email == '':
+        (valid, message) = validate({
+            'userIdEmail': {
+                'value': userId,
+                'validator': 'EMAIL'
+            }
+        })
+
+        if not valid:
+            email = "INVALID_FORMAT"
+        
+    return email 
+
+
 def create_subscription(body):
     response = STANDARD_JSON_RESPONSE
     subscription_table = dynamodb.Table(subscription_table_name)
     items = get_subscription_obj(body["eventName"], body["entityName"], body["entityId"])
+
+    #Lookup users email
+    emails = []
+    for subscriber in body["subscribers"]:
+        email = get_userProfile_Email(subscriber)
+        if email == "INVALID_FORMAT":
+            response['statusCode'] = 400
+            response['body'] = json.dumps({"message": f"Subscriber {subscriber} does not have a valid email to use."})
+            return response
+        else:
+            emails.append(email)
 
     if not items:
         subscription_table.put_item(
@@ -219,7 +263,7 @@ def create_subscription(body):
 
         if body["entityName"] == "Asset":
             logger.info("creating subscription")
-            create_sns_subscriptions(body["entityId"], body["subscribers"])
+            create_sns_subscriptions(body["entityId"], emails)
 
     else:
         existing_subscribers = [item["S"] for item in items["subscribers"]['L']]
@@ -229,7 +273,7 @@ def create_subscription(body):
             return response
         else:
             if body["entityName"] == "Asset":
-                create_sns_subscriptions(body["entityId"], body["subscribers"])
+                create_sns_subscriptions(body["entityId"], emails)
 
             subscription_table.update_item(
                 Key={
@@ -262,6 +306,26 @@ def update_subscription(body):
     deleted_subscribers = set(existing_subscribers) - set(new_subscribers)
     added_subscribers = set(new_subscribers) - set(existing_subscribers)
 
+    #Lookup users email
+    emailsAdded = []
+    emailsDeleted = []
+    for subscriber in added_subscribers:
+        email = get_userProfile_Email(subscriber)
+        if email == "INVALID_FORMAT":
+            response['statusCode'] = 400
+            response['body'] = json.dumps({"message": f"Subscriber {subscriber} does not have a valid email to use."})
+            return response
+        else:
+            emailsAdded.append(email)
+    for subscriber in deleted_subscribers:
+        email = get_userProfile_Email(subscriber)
+        if email == "INVALID_FORMAT":
+            response['statusCode'] = 400
+            response['body'] = json.dumps({"message": f"Subscriber {subscriber} does not have a valid email to use."})
+            return response
+        else:
+            emailsDeleted.append(email)
+
     subscription_table.update_item(
         Key={
             'eventName': body["eventName"],
@@ -274,8 +338,8 @@ def update_subscription(body):
     )
 
     if body["entityName"] == "Asset":
-        create_sns_subscriptions(body["entityId"], list(added_subscribers))
-        delete_sns_subscriptions(body["entityId"], list(deleted_subscribers), delete_sns=False)
+        create_sns_subscriptions(body["entityId"], list(emailsAdded))
+        delete_sns_subscriptions(body["entityId"], list(emailsDeleted), delete_sns=False)
 
     response['statusCode'] = 200
     response['body'] = json.dumps({"message": "success"})
@@ -301,9 +365,10 @@ def delete_subscription(body):
             response['statusCode'] = 500
             response['body'] = json.dumps({"message": "An unexpected error occurred while executing the request"})
         return response
+    
 
     if body["entityName"] == "Asset":
-        delete_sns_subscriptions(body["entityId"], body["subscribers"], delete_sns=True)
+        delete_sns_subscriptions(body["entityId"], None, delete_sns=True)
 
     response['statusCode'] = 200
     response['body'] = json.dumps({"message": "success"})
@@ -361,7 +426,7 @@ def lambda_handler(event, context):
             },
             'subscribers': {
                 'value': event['body']['subscribers'],
-                'validator': 'EMAIL_ARRAY'
+                'validator': 'USERID_ARRAY'
             }
         })