+++ infra/lib/core-stack.ts
@@ -46,10 +46,17 @@ export class CoreVAMSStack extends cdk.Stack {
     constructor(scope: Construct, id: string, props: EnvProps) {
         super(scope, id, { ...props, crossRegionReferences: true });
 
+        const adminUserId = new cdk.CfnParameter(this, "adminUserId", {
+            type: "String",
+            description:
+                "Admin User ID for login",
+            default: props.config.app.adminUserId,
+        });
+
         const adminEmailAddress = new cdk.CfnParameter(this, "adminEmailAddress", {
             type: "String",
             description:
-                "Email address for login and where your password is sent to. You will be sent a temporary password to authenticate to Cognito.",
+                "Admin Email address for login and where your password is sent to. You will be sent a temporary password to authenticate to Cognito.",
             default: props.config.app.adminEmailAddress,
         });