+++ infra/lib/nestedStacks/auth/authBuilder-nestedStack.ts
@@ -25,6 +25,13 @@ import * as ec2 from "aws-cdk-lib/aws-ec2";
 import * as iam from "aws-cdk-lib/aws-iam";
 import { samlSettings } from "../../../config/saml-config";
 import { Service } from "../../../lib/helper/service-helper";
+import {
+    AwsCustomResource,
+    AwsSdkCall,
+    AwsCustomResourcePolicy,
+    PhysicalResourceId,
+    PhysicalResourceIdReference,
+} from "aws-cdk-lib/custom-resources";
 
 export interface authResources {
     roles: {
@@ -111,6 +118,11 @@ export class AuthBuilderNestedStack extends NestedStack {
                 new iam.PolicyStatement({
                     effect: iam.Effect.ALLOW,
                     actions: ["dynamodb:PutItem"],
+                    resources: [props.storageResources.dynamo.userStorageTable.tableArn],
+                }),
+                new iam.PolicyStatement({
+                    effect: iam.Effect.ALLOW,
+                    actions: ["dynamodb:PutItem"],
                     resources: [props.storageResources.dynamo.authEntitiesStorageTable.tableArn],
                 }),
             ],
@@ -138,7 +150,38 @@ export class AuthBuilderNestedStack extends NestedStack {
             );
         }
 
-        //Deploy DynamoDB defaults
+        //Deploy DynamoDB admin user login profile
+        const awsSdkCallUserAdmin: AwsSdkCall = {
+            service: "DynamoDB",
+            action: "putItem",
+            parameters: {
+                TableName: props.storageResources.dynamo.userStorageTable.tableName,
+                Item: {
+                    userId: {
+                        S: props.config.app.adminUserId,
+                    },
+                    email: {
+                        S: props.config.app.adminEmailAddress,
+                    },
+                    createdOn: {
+                        S: new Date().toISOString(),
+                    }
+                },
+                //ConditionExpression: "attribute_not_exists(userId)",
+            },
+            physicalResourceId: PhysicalResourceId.of(
+                props.storageResources.dynamo.userStorageTable.tableName +
+                    `_admin_initialization`
+            ),
+        };
+
+        new AwsCustomResource(this, `userStorageTable_admin_CustomResource`, {
+            onCreate: awsSdkCallUserAdmin,
+            onUpdate: awsSdkCallUserAdmin,
+            role: authDefaultCustomResourceRole,
+        });
+
+        //Deploy DynamoDB for roles/constraints defaults
         const dynamoDbAuthDefaultsAdmin = new DynamoDbAuthDefaultsAdminConstructStack(
             this,
             "DynamoDBAuthDefaultsAdmin",