+++ infra/lib/nestedStacks/auth/constructs/dynamodb-authdefaults-admin-construct.ts
@@ -78,6 +78,7 @@ export class DynamoDbAuthDefaultsAdminConstructStack extends Construct {
             role: props.customResourceRole,
         });
 
+
         const awsSdkCallUserRolesAdmin: AwsSdkCall = {
             service: "DynamoDB",
             action: "putItem",
@@ -85,7 +86,7 @@ export class DynamoDbAuthDefaultsAdminConstructStack extends Construct {
                 TableName: props.storageResources.dynamo.userRolesStorageTable.tableName,
                 Item: {
                     userId: {
-                        S: props.config.app.adminEmailAddress,
+                        S: props.config.app.adminUserId,
                     },
                     roleName: {
                         S: roleName,