+++ infra/lib/nestedStacks/search/constructs/opensearch-provisioned.ts
@@ -196,6 +196,7 @@ export class OpensearchProvisionedConstruct extends Construct {
                     externalModules: ["aws-sdk"],
                 },
                 runtime: LAMBDA_NODE_RUNTIME,
+                timeout: cdk.Duration.seconds(30),
                 vpc: props.vpc,
                 vpcSubnets: { subnets: props.subnets },
                 //Note: This schema deploy resource must run in the VPC in order to communicate with the AOS provisioned running in the VPC.